add required media type: moved gen functions to pdr.utils
