+++ python/nistoar/pdr/preserv/bagit/exceptions.py
@@ -27,3 +27,9 @@ class BadBagRequest(BagItException):
     """
     pass
 
+class ComponentNotFound(BadBagRequest):
+    """
+    an exception resulting from a request of a non-existent component
+    """
+    pass
+
