+++ python/tests/nistoar/pdr/preserv/bagger/test_midas.py
@@ -149,7 +149,7 @@ class TestMIDASMetadataBaggerMixed(test.TestCase):
         self.bagr.bagbldr.ediid = "gurn"
 
         destpath = os.path.join("trial3", "trial3a.json")
-        dlurl = "https://www.nist.gov/od/ds/gurn/"+destpath
+        dlurl = "https://data.nist.gov/od/ds/gurn/"+destpath
         dfile = os.path.join(self.upldir, self.midasid[32:], destpath)
         self.bagr.ensure_file_metadata(dfile, destpath)
